+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Incremental improvements and new preview features.
 
 ### New features:
 - Support having a dynamic *rootDir* property. Allow *apiUserId* and *effectiveUserId*. Attribute *rootDir* may start with *HOST_EVAL()*.
+- For a dynamic *rootDir*, create the resolved *rootDir* on the system host as needed during the credential registration process.
 - Add support for getSystem *resolve* query parameter. Options are *ALL*, *NONE*, *ROOT_DIR*, *EFFECTIVE_USER*.
 - Add attributes returned for getSystem: *isPublic*, *isDynamicEffectiveUser*, *isDynamicRootDir*.
 - Add endpoint for checking credentials.